day 6 fix: xor lists aren't possible in Go any more
